# improve Format() performance